All Operators' correctness is verified

+++ src/Operator/OperatorAffine.cpp
@@ -4,15 +4,15 @@ namespace MetaPB {
 namespace Operator {
 
 inline void OperatorAFFINE::execCPU(const CPU_TCB& cpuTCB) const noexcept {
-  float* src1 = (float*)cpuTCB.src1PageBase;
-  float* src2 = (float*)cpuTCB.src2PageBase;
-  float* dst = (float*)cpuTCB.dstPageBase;
+  int* src1 = (int*)cpuTCB.src1PageBase;
+  int* src2 = (int*)cpuTCB.src2PageBase;
+  int* dst =  (int*)cpuTCB.dstPageBase;
   uint32_t itemNum = cpuTCB.pageBlkCnt * pageBlkSize / sizeof(float);
 
   omp_set_num_threads(64);
 #pragma omp parallel for
   for (int i = 0; i < itemNum; i++) {
-    dst[i] = src1[i] * weight + src2[i];
+    dst[i] = src1[i] * (int)weight + src2[i];
   }
 }
 inline void OperatorAFFINE::execDPU(const DPU_TCB& dpuTCB) const noexcept {
